Stop audit trail TimerTasks waking between intervals
Both the TimerTasks for collection and preservation were set to wake 10 times within their intervals checking if it was time to do stuff and otherwise doing nothing. So waking 9 times with no purpose. No explaining comment as to why. Timer#scheduleAtFixedRate only schedules tasks on one thread, so scheduling more tasks than needed does not seem to have any benefit even if it was in case tasks were taking long to complete. The TimerTasks now only wake on their intervals.