jQuery migration to 3.6.1
NOTE: The plotting library "flot" has been tweaked to comply with the new jQuery-version, but has itself not been updated. An attempt was made at upgrading it, but this was shelved because of trouble with making the usage work as before (its repository is poorly maintained and errors are difficult to troubleshoot). Either way this would have little payoff.